Reorganize OAuth2 code
