+++ src/models/oauthclients.ts
@@ -1,0 +1,8 @@
+import { URL } from 'url';
+import {Scope, ClientId, ClientRegistry} from '../oauth-types';
+
+const OAuthClients = new Map();
+OAuthClients.set('hello',
+                  new ClientRegistry(new URL('http://127.0.0.1:31338/callback'), 'hello', 'readwrite'));
+
+export default OAuthClients;
